Relocated all files to follow restructured directories.
