+++ src/main/java/edit_timetable_use_case/AddCourseInteractor.java
@@ -1,8 +1,8 @@
 package edit_timetable_use_case;
 
 import entities.*;
-import retrieve_timetable_use_case.RetrieveTimetableInteractor;
-import retrieve_timetable_use_case.TimetableModel;
+import retrieve_timetable_use_case.application_business.RetrieveTimetableInteractor;
+import retrieve_timetable_use_case.application_business.TimetableModel;
 
 import java.util.ArrayList;
 import java.util.List;
