+++ src/main/java/retrieve_timetable_use_case/application_business/RetrieveTimetableInputBoundary.java
@@ -1,4 +1,4 @@
-package retrieve_timetable_use_case;
+package retrieve_timetable_use_case.application_business;
 
 import entities.Session;
 import entities.Timetable;
