+++ src/main/java/retrieve_timetable_use_case/interface_adapters/RetrieveTimetableController.java
@@ -1,4 +1,9 @@
-package retrieve_timetable_use_case;
+package retrieve_timetable_use_case.interface_adapters;
+
+import retrieve_timetable_use_case.application_business.RetrieveTimetableInputBoundary;
+import retrieve_timetable_use_case.application_business.RetrieveTimetableRequestModel;
+import retrieve_timetable_use_case.application_business.CourseModel;
+import retrieve_timetable_use_case.application_business.TimetableModel;
 
 /**
  * The controller used in the RetrieveTimetableUseCase.
