+++ src/test/java/retrieve_timetable_use_case/CourseModelTest.java
@@ -2,11 +2,13 @@ package retrieve_timetable_use_case;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import retrieve_timetable_use_case.application_business.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 /**
  * A test suite for the CourseModel class, primarily to confirm the correctness of BlockModel's equal method.
