+++ src/test/java/retrieve_timetable_use_case/SectionModelTest.java
@@ -2,10 +2,12 @@ package retrieve_timetable_use_case;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import retrieve_timetable_use_case.application_business.*;
 
 import java.util.ArrayList;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 /**
  * A test suite for the SectionModel class, primarily to confirm the correctness of its equal method.
