+++ src/test/java/retrieve_timetable_use_case/SessionModelTest.java
@@ -3,10 +3,13 @@ package retrieve_timetable_use_case;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import retrieve_timetable_use_case.application_business.*;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import static org.junit.jupiter.api.Assertions.*;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 /**
  * A test suite for the SessionModel class, primarily to confirm the correctness of its equal method.
